+++ pyobs/__init__.py
@@ -1,3 +1,3 @@
-from .version import __version__, version_tuple
+from .version import __version__, version
 
-__all__ = ["version", "__version__", "version_tuple"]
+__all__ = ["version", "__version__"]